Fix error snackbar display issue
Positioning of the snackbar element inside the settings draw caused display issues on computer when closing the drawer and on mobile. Moved the snackbar element out of the drawer and wrapped both elements in a Fragment element to fix the issue.

Also changed settings draw width (250 --> 220) and modified the checkbox layout.